[Game] WIP: Work with dungeons; - code for working with dungeons and Indun tables for the database from Genesis; - now each instance has its own Id for the team and for the personal dungeon; - Added handling of events for players and NPCs when working with dungeons; - Added check for the number of attempts to pass the dungeon; - after death in the dungeon, the character is transferred to Priests of Nui into dungeon;

+++ AAEmu.Game/Core/Managers/AuctionManager.cs
@@ -159,7 +159,7 @@ public class AuctionManager : Singleton<AuctionManager>
                     newMail.Send();
                 }
 
-                player.SubtractMoney(SlotType.Inventory, (int)auctionItem.DirectMoney);
+                player.SubtractMoney(SlotType.Inventory, auctionItem.DirectMoney);
                 RemoveAuctionItemSold(auctionItem, player.Name, auctionItem.DirectMoney);
             }
 
@@ -168,7 +168,7 @@ public class AuctionManager : Singleton<AuctionManager>
                 if ((auctionItem.BidderName != "") && (auctionItem.BidderId != 0)) //Send mail to old bidder. 
                 {
                     var moneyArray = new int[3];
-                    moneyArray[0] = (int)auctionItem.BidMoney;
+                    moneyArray[0] = auctionItem.BidMoney;
 
                     // TODO: Read this from saved data
                     var recalculatedFee = (auctionItem.DirectMoney * .01) * (auctionItem.Duration + 1);
@@ -185,7 +185,7 @@ public class AuctionManager : Singleton<AuctionManager>
                 auctionItem.BidWorldID = (byte)player.Transform.WorldId;
                 auctionItem.BidMoney = bidAmount;
 
-                player.SubtractMoney(SlotType.Inventory, (int)bidAmount);
+                player.SubtractMoney(SlotType.Inventory, bidAmount);
                 player.SendPacket(new SCAuctionBidPacket(auctionItem));
                 auctionItem.IsDirty = true;
             }
